Move shit out of MainUtil into more appropriate classes
