+++ Core/src/main/java/com/plotsquared/core/command/Buy.java
@@ -36,12 +36,11 @@ import com.plotsquared.core.plot.flag.PlotFlag;
 import com.plotsquared.core.plot.flag.implementations.PriceFlag;
 import com.plotsquared.core.util.EconHandler;
 import com.plotsquared.core.util.EventDispatcher;
-import com.plotsquared.core.util.MainUtil;
 import com.plotsquared.core.util.task.RunnableVal2;
 import com.plotsquared.core.util.task.RunnableVal3;
-import javax.annotation.Nullable;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 
@@ -72,7 +71,7 @@ public class Buy extends Command {
         final Plot plot;
         if (args.length != 0) {
             checkTrue(args.length == 1, Captions.COMMAND_SYNTAX, getUsage());
-            plot = check(MainUtil.getPlotFromString(player, args[0], true), null);
+            plot = check(Plot.getPlotFromString(player, args[0], true), null);
         } else {
             plot = check(player.getCurrentPlot(), Captions.NOT_IN_PLOT);
         }
