+++ Core/src/main/java/com/plotsquared/core/command/Copy.java
@@ -56,7 +56,7 @@ public class Copy extends SubCommand {
             Captions.COMMAND_SYNTAX.send(player, getUsage());
             return false;
         }
-        Plot plot2 = MainUtil.getPlotFromString(player, args[0], true);
+        Plot plot2 = Plot.getPlotFromString(player, args[0], true);
         if (plot2 == null) {
             return false;
         }
