+++ Core/src/main/java/com/plotsquared/core/command/Target.java
@@ -69,7 +69,7 @@ public class Target extends SubCommand {
                 MainUtil.sendMessage(player, Captions.FOUND_NO_PLOTS);
                 return false;
             }
-        } else if ((target = MainUtil.getPlotFromString(player, args[0], true)) == null) {
+        } else if ((target = Plot.getPlotFromString(player, args[0], true)) == null) {
             return false;
         }
         target.getCenter(player::setCompassTarget);
