+++ Core/src/main/java/com/plotsquared/core/generator/HybridPlotManager.java
@@ -37,9 +37,10 @@ import com.plotsquared.core.plot.PlotId;
 import com.plotsquared.core.queue.LocalBlockQueue;
 import com.plotsquared.core.util.ChunkManager;
 import com.plotsquared.core.util.FileBytes;
-import com.plotsquared.core.util.MainUtil;
+import com.plotsquared.core.util.FileUtils;
 import com.plotsquared.core.util.MathMan;
 import com.plotsquared.core.util.RegionManager;
+import com.plotsquared.core.util.WorldUtil;
 import com.plotsquared.core.util.task.RunnableVal;
 import com.sk89q.worldedit.function.pattern.Pattern;
 import com.sk89q.worldedit.world.biome.BiomeType;
@@ -47,12 +48,13 @@ import com.sk89q.worldedit.world.block.BaseBlock;
 import com.sk89q.worldedit.world.block.BlockState;
 import com.sk89q.worldedit.world.block.BlockTypes;
 import lombok.Getter;
-import javax.annotation.Nonnull;
 
+import javax.annotation.Nonnull;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.util.HashSet;
+import java.util.Objects;
 
 public class HybridPlotManager extends ClassicPlotManager {
 
@@ -77,7 +79,7 @@ public class HybridPlotManager extends ClassicPlotManager {
                 .getWorldName() + File.separator;
         try {
             File sideRoad =
-                MainUtil.getFile(PlotSquared.platform().getDirectory(), dir + "sideroad.schem");
+                FileUtils.getFile(PlotSquared.platform().getDirectory(), dir + "sideroad.schem");
             String newDir = "schematics" + File.separator + "GEN_ROAD_SCHEMATIC" + File.separator
                 + "__TEMP_DIR__" + File.separator;
             if (sideRoad.exists()) {
@@ -85,12 +87,12 @@ public class HybridPlotManager extends ClassicPlotManager {
                     Files.readAllBytes(sideRoad.toPath())));
             }
             File intersection =
-                MainUtil.getFile(PlotSquared.platform().getDirectory(), dir + "intersection.schem");
+                FileUtils.getFile(PlotSquared.platform().getDirectory(), dir + "intersection.schem");
             if (intersection.exists()) {
                 files.add(new FileBytes(newDir + "intersection.schem",
                     Files.readAllBytes(intersection.toPath())));
             }
-            File plot = MainUtil.getFile(PlotSquared.platform().getDirectory(), dir + "plot.schem");
+            File plot = FileUtils.getFile(PlotSquared.platform().getDirectory(), dir + "plot.schem");
             if (plot.exists()) {
                 files.add(new FileBytes(newDir + "plot.schem", Files.readAllBytes(plot.toPath())));
             }
@@ -109,7 +111,7 @@ public class HybridPlotManager extends ClassicPlotManager {
         Location pos1 = Location.at(hybridPlotWorld.getWorldName(), top.getX() + 1, 0, bot.getZ() - 1);
         Location pos2 = Location.at(hybridPlotWorld.getWorldName(), bot.getX(),
             Math.min(getWorldHeight(), 255), top.getZ() + 1);
-        MainUtil.resetBiome(hybridPlotWorld, pos1, pos2);
+        this.resetBiome(hybridPlotWorld, pos1, pos2);
         if (!hybridPlotWorld.ROAD_SCHEMATIC_ENABLED) {
             return true;
         }
@@ -117,6 +119,17 @@ public class HybridPlotManager extends ClassicPlotManager {
         createSchemAbs(queue, pos1, pos2, true);
         queue.enqueue();
         return true;
+    }
+
+    private void resetBiome(@Nonnull final HybridPlotWorld hybridPlotWorld, @Nonnull final Location pos1,
+        @Nonnull final Location pos2) {
+        BiomeType biome = hybridPlotWorld.getPlotBiome();
+        if (!Objects.equals(PlotSquared.platform().getWorldUtil()
+            .getBiomeSynchronous(hybridPlotWorld.getWorldName(), (pos1.getX() + pos2.getX()) / 2,
+                (pos1.getZ() + pos2.getZ()) / 2), biome)) {
+            WorldUtil.setBiome(hybridPlotWorld.getWorldName(), pos1.getX(), pos1.getZ(), pos2.getX(), pos2.getZ(),
+                    biome);
+        }
     }
 
     private void createSchemAbs(LocalBlockQueue queue, Location pos1, Location pos2,
@@ -170,7 +183,7 @@ public class HybridPlotManager extends ClassicPlotManager {
         Location pos1 = Location.at(hybridPlotWorld.getWorldName(), bot.getX() - 1, 0, top.getZ() + 1);
         Location pos2 = Location.at(hybridPlotWorld.getWorldName(), top.getX() + 1,
             Math.min(getWorldHeight(), 255), bot.getZ());
-        MainUtil.resetBiome(hybridPlotWorld, pos1, pos2);
+        this.resetBiome(hybridPlotWorld, pos1, pos2);
         if (!hybridPlotWorld.ROAD_SCHEMATIC_ENABLED) {
             return true;
         }
@@ -238,7 +251,7 @@ public class HybridPlotManager extends ClassicPlotManager {
                 }
                 /* Otherwise we need to set each component, as we don't want to regenerate the road or other plots that share the same chunk.*/
                 // Set the biome
-                MainUtil.setBiome(world, value[2], value[3], value[4], value[5], biome);
+                WorldUtil.setBiome(world, value[2], value[3], value[4], value[5], biome);
                 // These two locations are for each component (e.g. bedrock, main block, floor, air)
                 Location bot = Location.at(world, value[2], 0, value[3]);
                 Location top = Location.at(world, value[4], 1, value[5]);
