+++ Core/src/main/java/com/plotsquared/core/listener/PlotListener.java
@@ -58,16 +58,16 @@ import com.plotsquared.core.plot.flag.types.TimedFlag;
 import com.plotsquared.core.util.EventDispatcher;
 import com.plotsquared.core.util.MainUtil;
 import com.plotsquared.core.util.Permissions;
+import com.plotsquared.core.util.PlayerManager;
 import com.plotsquared.core.util.StringMan;
-import com.plotsquared.core.util.task.RunnableVal;
 import com.plotsquared.core.util.task.TaskManager;
 import com.plotsquared.core.util.task.TaskTime;
 import com.sk89q.worldedit.world.gamemode.GameMode;
 import com.sk89q.worldedit.world.gamemode.GameModes;
 import com.sk89q.worldedit.world.item.ItemType;
 import com.sk89q.worldedit.world.item.ItemTypes;
-import javax.annotation.Nullable;
 
+import javax.annotation.Nullable;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
@@ -155,12 +155,8 @@ public class PlotListener {
 
             final String greeting = plot.getFlag(GreetingFlag.class);
             if (!greeting.isEmpty()) {
-                MainUtil.format(Captions.PREFIX_GREETING.getTranslated() + greeting, plot, player,
-                    false, new RunnableVal<String>() {
-                        @Override public void run(String value) {
-                            MainUtil.sendMessage(player, value);
-                        }
-                    });
+                plot.format(Captions.PREFIX_GREETING.getTranslated() + greeting, player, false)
+                    .thenAcceptAsync(player::sendMessage);
             }
 
             if (plot.getFlag(NotifyEnterFlag.class)) {
@@ -271,7 +267,7 @@ public class PlotListener {
                             replacements.put("%world%", plot.getArea().toString());
                             replacements.put("%greeting%", greeting);
                             replacements.put("%alias", plot.toString());
-                            replacements.put("%s", MainUtil.getName(plot.getOwner()));
+                            replacements.put("%s", PlayerManager.getName(plot.getOwner()));
                             String main = StringMan
                                 .replaceFromMap(Captions.TITLE_ENTERED_PLOT.getTranslated(),
                                     replacements);
@@ -331,12 +327,8 @@ public class PlotListener {
 
             final String farewell = plot.getFlag(FarewellFlag.class);
             if (!farewell.isEmpty()) {
-                MainUtil.format(Captions.PREFIX_FAREWELL.getTranslated() + farewell, plot, player,
-                    false, new RunnableVal<String>() {
-                        @Override public void run(String value) {
-                            MainUtil.sendMessage(player, value);
-                        }
-                    });
+                plot.format(Captions.PREFIX_FAREWELL.getTranslated() + farewell, player, false)
+                    .thenAcceptAsync(player::sendMessage);
             }
 
             if (plot.getFlag(NotifyLeaveFlag.class)) {
