+++ Core/src/main/java/com/plotsquared/core/plot/Plot.java
@@ -26,9 +26,12 @@
 package com.plotsquared.core.plot;
 
 import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import com.google.inject.Inject;
 import com.plotsquared.core.PlotSquared;
+import com.plotsquared.core.command.Like;
+import com.plotsquared.core.configuration.CaptionUtility;
 import com.plotsquared.core.configuration.Captions;
 import com.plotsquared.core.configuration.ConfigurationUtil;
 import com.plotsquared.core.configuration.Settings;
@@ -45,6 +48,7 @@ import com.plotsquared.core.location.BlockLoc;
 import com.plotsquared.core.location.Direction;
 import com.plotsquared.core.location.Location;
 import com.plotsquared.core.location.PlotLoc;
+import com.plotsquared.core.player.ConsolePlayer;
 import com.plotsquared.core.player.PlotPlayer;
 import com.plotsquared.core.plot.comment.PlotComment;
 import com.plotsquared.core.plot.expiration.ExpireManager;
@@ -53,7 +57,10 @@ import com.plotsquared.core.plot.flag.FlagContainer;
 import com.plotsquared.core.plot.flag.GlobalFlagContainer;
 import com.plotsquared.core.plot.flag.InternalFlag;
 import com.plotsquared.core.plot.flag.PlotFlag;
+import com.plotsquared.core.plot.flag.implementations.DescriptionFlag;
 import com.plotsquared.core.plot.flag.implementations.KeepFlag;
+import com.plotsquared.core.plot.flag.implementations.ServerPlotFlag;
+import com.plotsquared.core.plot.flag.types.DoubleFlag;
 import com.plotsquared.core.plot.schematic.Schematic;
 import com.plotsquared.core.queue.GlobalBlockQueue;
 import com.plotsquared.core.queue.LocalBlockQueue;
@@ -61,9 +68,13 @@ import com.plotsquared.core.util.EventDispatcher;
 import com.plotsquared.core.util.MainUtil;
 import com.plotsquared.core.util.MathMan;
 import com.plotsquared.core.util.Permissions;
+import com.plotsquared.core.util.PlayerManager;
 import com.plotsquared.core.util.RegionManager;
+import com.plotsquared.core.util.RegionUtil;
 import com.plotsquared.core.util.SchematicHandler;
+import com.plotsquared.core.util.TimeUtil;
 import com.plotsquared.core.util.WorldUtil;
+import com.plotsquared.core.util.query.PlotQuery;
 import com.plotsquared.core.util.task.RunnableVal;
 import com.plotsquared.core.util.task.TaskManager;
 import com.plotsquared.core.util.task.TaskTime;
@@ -86,6 +97,7 @@ import java.awt.geom.PathIterator;
 import java.awt.geom.Rectangle2D;
 import java.io.File;
 import java.net.URL;
+import java.text.DecimalFormat;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -124,23 +136,17 @@ import static com.plotsquared.core.util.entity.EntityCategories.CAP_VEHICLE;
  */
 public class Plot {
 
-    private static final Logger logger = LoggerFactory.getLogger("P2/" + Plot.class.getSimpleName());
-
     public static final int MAX_HEIGHT = 256;
+
+    private static final Logger logger = LoggerFactory.getLogger("P2/" + Plot.class.getSimpleName());
+    private static final DecimalFormat FLAG_DECIMAL_FORMAT = new DecimalFormat("0");
+
+    static {
+        FLAG_DECIMAL_FORMAT.setMaximumFractionDigits(340);
+    }
 
     private static Set<Plot> connected_cache;
     private static Set<CuboidRegion> regions_cache;
-
-    @Nonnull private PlotId id;
-
-    // These will be injected
-    @Inject private EventDispatcher eventDispatcher;
-    @Inject private PlotListener plotListener;
-    @Inject private RegionManager regionManager;
-    @Inject private GlobalBlockQueue blockQueue;
-    @Inject private WorldUtil worldUtil;
-    @Inject private SchematicHandler schematicHandler;
-    @Inject @ImpromptuPipeline private UUIDPipeline impromptuPipeline;
 
     /**
      * Plot flag container
@@ -150,7 +156,6 @@ public class Plot {
      * Has the plot changed since the last save cycle?
      */
     public boolean countsTowardsMax = true;
-
     /**
      * Represents whatever the database manager needs it to: <br>
      * - A value of -1 usually indicates the plot will not be stored in the DB<br>
@@ -159,6 +164,15 @@ public class Plot {
      * @deprecated magical
      */
     @Deprecated public int temp;
+    @Nonnull private PlotId id;
+    // These will be injected
+    @Inject private EventDispatcher eventDispatcher;
+    @Inject private PlotListener plotListener;
+    @Inject private RegionManager regionManager;
+    @Inject private GlobalBlockQueue blockQueue;
+    @Inject private WorldUtil worldUtil;
+    @Inject private SchematicHandler schematicHandler;
+    @Inject @ImpromptuPipeline private UUIDPipeline impromptuPipeline;
     /**
      * plot owner
      * (Merged plots can have multiple owners)
@@ -285,6 +299,72 @@ public class Plot {
     }
 
     /**
+     * Get the plot from a string.
+     *
+     * @param player  Provides a context for what world to search in. Prefixing the term with 'world_name;' will override this context.
+     * @param arg     The search term
+     * @param message If a message should be sent to the player if a plot cannot be found
+     * @return The plot if only 1 result is found, or null
+     */
+    @Nullable public static Plot getPlotFromString(PlotPlayer<?> player, String arg,
+        boolean message) {
+        if (arg == null) {
+            if (player == null) {
+                if (message) {
+                    logger.info("[P2] No plot area string was supplied");
+                }
+                return null;
+            }
+            return player.getCurrentPlot();
+        }
+        PlotArea area;
+        if (player != null) {
+            area = PlotSquared.get().getPlotAreaManager().getPlotAreaByString(arg);
+            if (area == null) {
+                area = player.getApplicablePlotArea();
+            }
+        } else {
+            area = ConsolePlayer.getConsole().getApplicablePlotArea();
+        }
+        String[] split = arg.split(";|,");
+        PlotId id;
+        if (split.length == 4) {
+            area = PlotSquared.get().getPlotAreaManager()
+                .getPlotAreaByString(split[0] + ';' + split[1]);
+            id = PlotId.fromString(split[2] + ';' + split[3]);
+        } else if (split.length == 3) {
+            area = PlotSquared.get().getPlotAreaManager().getPlotAreaByString(split[0]);
+            id = PlotId.fromString(split[1] + ';' + split[2]);
+        } else if (split.length == 2) {
+            id = PlotId.fromString(arg);
+        } else {
+            Collection<Plot> plots;
+            if (area == null) {
+                plots = PlotQuery.newQuery().allPlots().asList();
+            } else {
+                plots = area.getPlots();
+            }
+            for (Plot p : plots) {
+                String name = p.getAlias();
+                if (!name.isEmpty() && name.equalsIgnoreCase(arg)) {
+                    return p;
+                }
+            }
+            if (message) {
+                MainUtil.sendMessage(player, Captions.NOT_VALID_PLOT_ID);
+            }
+            return null;
+        }
+        if (area == null) {
+            if (message) {
+                MainUtil.sendMessage(player, Captions.NOT_VALID_PLOT_WORLD);
+            }
+            return null;
+        }
+        return area.getPlotAbs(id);
+    }
+
+    /**
      * Gets a plot from a string e.g. [area];[id]
      *
      * @param defaultArea if no area is specified
@@ -327,6 +407,13 @@ public class Plot {
             return pa.getPlot(location);
         }
         return null;
+    }
+
+    @Nonnull private static Location[] getCorners(@Nonnull final String world,
+        @Nonnull final CuboidRegion region) {
+        final BlockVector3 min = region.getMinimumPoint();
+        final BlockVector3 max = region.getMaximumPoint();
+        return new Location[] {Location.at(world, min), Location.at(world, max)};
     }
 
     /**
@@ -471,7 +558,7 @@ public class Plot {
      * @see #getOwnerAbs() getOwnerAbs() to get the owner as stored in the database
      */
     public UUID getOwner() {
-        if (MainUtil.isServerOwned(this)) {
+        if (this.getFlag(ServerPlotFlag.class)) {
             return DBFunc.SERVER;
         }
         return this.getOwnerAbs();
@@ -576,6 +663,15 @@ public class Plot {
      */
     @Nonnull public PlotId getId() {
         return this.id;
+    }
+
+    /**
+     * Change the plot ID
+     *
+     * @param id new plot ID
+     */
+    public void setId(@Nonnull final PlotId id) {
+        this.id = id;
     }
 
     /**
@@ -961,7 +1057,7 @@ public class Plot {
                 if (queue.isEmpty()) {
                     Runnable run = () -> {
                         for (CuboidRegion region : regions) {
-                            Location[] corners = MainUtil.getCorners(getWorldName(), region);
+                            Location[] corners = getCorners(getWorldName(), region);
                             regionManager.clearAllEntities(corners[0], corners[1]);
                         }
                         TaskManager.runTask(whenDone);
@@ -1071,7 +1167,7 @@ public class Plot {
             blockQueue.addEmptyTask(() -> {
                 TaskManager.runTaskAsync(() -> {
                     for (Plot current : plots) {
-                        current.setSign(MainUtil.getName(current.getOwnerAbs()));
+                        current.setSign(PlayerManager.getName(current.getOwnerAbs()));
                     }
                 });
             });
@@ -1859,6 +1955,8 @@ public class Plot {
         return this.setComponent(component, parsed.toPattern());
     }
 
+    //TODO Better documentation needed.
+
     /**
      * Retrieve the biome of the plot.
      */
@@ -1866,6 +1964,8 @@ public class Plot {
         this.getCenter(location -> this.worldUtil
             .getBiome(location.getWorldName(), location.getX(), location.getZ(), result));
     }
+
+    //TODO Better documentation needed.
 
     /**
      * @deprecated May cause synchronous chunk loads
@@ -1876,16 +1976,12 @@ public class Plot {
             .getBiomeSynchronous(location.getWorldName(), location.getX(), location.getZ());
     }
 
-    //TODO Better documentation needed.
-
     /**
      * Returns the top location for the plot.
      */
     public Location getTopAbs() {
         return this.getManager().getPlotTopLocAbs(this.id).withWorld(this.getWorldName());
     }
-
-    //TODO Better documentation needed.
 
     /**
      * Returns the bottom location for the plot.
@@ -2001,7 +2097,7 @@ public class Plot {
         if (!this.isMerged()) {
             return new Location[] {this.getBottomAbs(), this.getTopAbs()};
         }
-        return MainUtil.getCorners(this.getWorldName(), this.getRegions());
+        return RegionUtil.getCorners(this.getWorldName(), this.getRegions());
     }
 
     /**
@@ -2077,7 +2173,6 @@ public class Plot {
         }
         return this.area + ";" + this.id.toString();
     }
-
 
     /**
      * Remove a denied player (use DBFunc as well)<br>
@@ -2179,8 +2274,8 @@ public class Plot {
                     }
                 } else {
                     TaskManager.runTaskAsync(() -> {
-                        String name = Plot.this.id + "," + Plot.this.area + ',' + MainUtil
-                            .getName(Plot.this.getOwnerAbs());
+                        String name = Plot.this.id + "," + Plot.this.area + ',' +
+                            PlayerManager.getName(Plot.this.getOwnerAbs());
                         boolean result = schematicHandler.save(value,
                             Settings.Paths.SCHEMATICS + File.separator + name + ".schem");
                         if (whenDone != null) {
@@ -2754,8 +2849,9 @@ public class Plot {
             PlotId top = PlotId.of(current.getId().getX(), current.getId().getY());
             while (merge) {
                 merge = false;
-                ArrayList<PlotId> ids = MainUtil.getPlotSelectionIds(PlotId.of(bot.getX(), bot.getY() - 1),
-                    PlotId.of(top.getX(), bot.getY() - 1));
+                List<PlotId> ids = Lists.newArrayList((Iterable<? extends PlotId>)
+                    PlotId.PlotRangeIterator.range(PlotId.of(bot.getX(), bot.getY() - 1),
+                        PlotId.of(top.getX(), bot.getY() - 1)));
                 boolean tmp = true;
                 for (PlotId id : ids) {
                     Plot plot = this.area.getPlotAbs(id);
@@ -2768,8 +2864,9 @@ public class Plot {
                     merge = true;
                     bot = PlotId.of(bot.getX(), bot.getY() - 1);
                 }
-                ids = MainUtil.getPlotSelectionIds(PlotId.of(top.getX() + 1, bot.getY()),
-                    PlotId.of(top.getX() + 1, top.getY()));
+                ids = Lists.newArrayList((Iterable<? extends PlotId>)
+                    PlotId.PlotRangeIterator.range(PlotId.of(top.getX() + 1, bot.getY()),
+                        PlotId.of(top.getX() + 1, top.getY())));
                 tmp = true;
                 for (PlotId id : ids) {
                     Plot plot = this.area.getPlotAbs(id);
@@ -2782,8 +2879,9 @@ public class Plot {
                     merge = true;
                     top = PlotId.of(top.getX() + 1, top.getY());
                 }
-                ids = MainUtil.getPlotSelectionIds(PlotId.of(bot.getX(), top.getY() + 1),
-                    PlotId.of(top.getX(), top.getY() + 1));
+                ids = Lists.newArrayList((Iterable<? extends PlotId>)
+                    PlotId.PlotRangeIterator.range(PlotId.of(bot.getX(), top.getY() + 1),
+                        PlotId.of(top.getX(), top.getY() + 1)));
                 tmp = true;
                 for (PlotId id : ids) {
                     Plot plot = this.area.getPlotAbs(id);
@@ -2796,8 +2894,9 @@ public class Plot {
                     merge = true;
                     top = PlotId.of(top.getX(), top.getY() + 1);
                 }
-                ids = MainUtil.getPlotSelectionIds(PlotId.of(bot.getX() - 1, bot.getY()),
-                    PlotId.of(bot.getX() - 1, top.getY()));
+                ids = Lists.newArrayList((Iterable<? extends PlotId>)
+                    PlotId.PlotRangeIterator.range(PlotId.of(bot.getX() - 1, bot.getY()),
+                        PlotId.of(bot.getX() - 1, top.getY())));
                 tmp = true;
                 for (PlotId id : ids) {
                     Plot plot = this.area.getPlotAbs(id);
@@ -2813,7 +2912,8 @@ public class Plot {
             }
             Location gtopabs = this.area.getPlotAbs(top).getTopAbs();
             Location gbotabs = this.area.getPlotAbs(bot).getBottomAbs();
-            visited.addAll(MainUtil.getPlotSelectionIds(bot, top));
+            visited.addAll(Lists.newArrayList((Iterable<? extends PlotId>)
+                PlotId.PlotRangeIterator.range(bot, top)));
             for (int x = bot.getX(); x <= top.getX(); x++) {
                 Plot plot = this.area.getPlotAbs(PlotId.of(x, top.getY()));
                 if (plot.getMerged(Direction.SOUTH)) {
@@ -3194,7 +3294,7 @@ public class Plot {
                             TaskManager.runTask(whenDone);
                         } else {
                             CuboidRegion region = regions.poll();
-                            Location[] corners = MainUtil.getCorners(getWorldName(), region);
+                            Location[] corners = getCorners(getWorldName(), region);
                             Location pos1 = corners[0];
                             Location pos2 = corners[1];
                             Location pos3 = pos1.add(offsetX, 0, offsetZ).withWorld(destination.getWorldName());
@@ -3226,7 +3326,7 @@ public class Plot {
                         }
                         final Runnable task = this;
                         CuboidRegion region = regions.poll();
-                        Location[] corners = MainUtil.getCorners(getWorldName(), region);
+                        Location[] corners = getCorners(getWorldName(), region);
                         final Location pos1 = corners[0];
                         final Location pos2 = corners[1];
                         Location newPos = pos1.add(offsetX, 0, offsetZ).withWorld(destination.getWorldName());
@@ -3320,7 +3420,7 @@ public class Plot {
                     return;
                 }
                 CuboidRegion region = regions.poll();
-                Location[] corners = MainUtil.getCorners(getWorldName(), region);
+                Location[] corners = getCorners(getWorldName(), region);
                 Location pos1 = corners[0];
                 Location pos2 = corners[1];
                 Location newPos = pos1 .add(offsetX, 0, offsetZ).withWorld(destination.getWorldName());
@@ -3381,13 +3481,158 @@ public class Plot {
         return FlagContainer.<T, V>castUnsafe(flagInstance).getValue();
     }
 
+    public CompletableFuture<String> format(final String iInfo, PlotPlayer<?> player, final boolean full) {
+        final CompletableFuture<String> future = new CompletableFuture<>();
+        int num = this.getConnectedPlots().size();
+        String alias = !this.getAlias().isEmpty() ? this.getAlias() : Captions.NONE.getTranslated();
+        Location bot = this.getCorners()[0];
+        PlotSquared.platform().getWorldUtil()
+            .getBiome(this.getWorldName(), bot.getX(), bot.getZ(), biome -> {
+                String info = iInfo;
+                String trusted = PlayerManager.getPlayerList(this.getTrusted());
+                String members = PlayerManager.getPlayerList(this.getMembers());
+                String denied = PlayerManager.getPlayerList(this.getDenied());
+                String seen;
+                if (Settings.Enabled_Components.PLOT_EXPIRY && ExpireManager.IMP != null) {
+                    if (this.isOnline()) {
+                        seen = Captions.NOW.getTranslated();
+                    } else {
+                        int time = (int) (ExpireManager.IMP.getAge(this) / 1000);
+                        if (time != 0) {
+                            seen = TimeUtil.secToTime(time);
+                        } else {
+                            seen = Captions.UNKNOWN.getTranslated();
+                        }
+                    }
+                } else {
+                    seen = Captions.NEVER.getTranslated();
+                }
+
+                String description = this.getFlag(DescriptionFlag.class);
+                if (description.isEmpty()) {
+                    description = Captions.PLOT_NO_DESCRIPTION.getTranslated();
+                }
+
+                StringBuilder flags = new StringBuilder();
+                Collection<PlotFlag<?, ?>> flagCollection = this.getApplicableFlags(true);
+                if (flagCollection.isEmpty()) {
+                    flags.append(Captions.NONE.getTranslated());
+                } else {
+                    String prefix = " ";
+                    for (final PlotFlag<?, ?> flag : flagCollection) {
+                        Object value;
+                        if (flag instanceof DoubleFlag && !Settings.General.SCIENTIFIC) {
+                            value = FLAG_DECIMAL_FORMAT.format(flag.getValue());
+                        } else {
+                            value = flag.toString();
+                        }
+                        flags.append(prefix).append(CaptionUtility
+                            .format(player, Captions.PLOT_FLAG_LIST.getTranslated(), flag.getName(),
+                                CaptionUtility.formatRaw(player, value.toString(), "")));
+                        prefix = ", ";
+                    }
+                }
+                boolean build = this.isAdded(player.getUUID());
+                String owner = this.getOwners().isEmpty() ? "unowned" : PlayerManager.getPlayerList(this.getOwners());
+                if (this.getArea() != null) {
+                    info = info.replace("%area%",
+                        this.getArea().getWorldName() + (this.getArea().getId() == null ?
+                            "" :
+                            "(" + this.getArea().getId() + ")"));
+                } else {
+                    info = info.replace("%area%", Captions.NONE.getTranslated());
+                }
+                info = info.replace("%id%", this.getId().toString());
+                info = info.replace("%alias%", alias);
+                info = info.replace("%num%", String.valueOf(num));
+                info = info.replace("%desc%", description);
+                info = info.replace("%biome%", biome.toString().toLowerCase());
+                info = info.replace("%owner%", owner);
+                info = info.replace("%members%", members);
+                info = info.replace("%player%", player.getName());
+                info = info.replace("%trusted%", trusted);
+                info = info.replace("%helpers%", members);
+                info = info.replace("%denied%", denied);
+                info = info.replace("%seen%", seen);
+                info = info.replace("%flags%", flags);
+                info = info.replace("%build%", String.valueOf(build));
+                if (info.contains("%rating%")) {
+                    final String newInfo = info;
+                    TaskManager.runTaskAsync(() -> {
+                        String info1;
+                        if (Settings.Ratings.USE_LIKES) {
+                            info1 = newInfo.replaceAll("%rating%",
+                                String.format("%.0f%%", Like.getLikesPercentage(this) * 100D));
+                        } else {
+                            int max = 10;
+                            if (Settings.Ratings.CATEGORIES != null && !Settings.Ratings.CATEGORIES
+                                .isEmpty()) {
+                                max = 8;
+                            }
+                            if (full && Settings.Ratings.CATEGORIES != null
+                                && Settings.Ratings.CATEGORIES.size() > 1) {
+                                double[] ratings = this.getAverageRatings();
+                                String rating = "";
+                                String prefix = "";
+                                for (int i = 0; i < ratings.length; i++) {
+                                    rating +=
+                                        prefix + Settings.Ratings.CATEGORIES.get(i) + '=' + String
+                                            .format("%.1f", ratings[i]);
+                                    prefix = ",";
+                                }
+                                info1 = newInfo.replaceAll("%rating%", rating);
+                            } else {
+                                info1 = newInfo.replaceAll("%rating%",
+                                    String.format("%.1f", this.getAverageRating()) + '/' + max);
+                            }
+                        }
+                        future.complete(info1);
+                    });
+                    return;
+                }
+                future.complete(info);
+            });
+        return future;
+    }
+
     /**
-     * Change the plot ID
+     * If rating categories are enabled, get the average rating by category.<br>
+     * - The index corresponds to the index of the category in the config
      *
-     * @param id new plot ID
+     * @return Average ratings in each category
      */
-    public void setId(@Nonnull final PlotId id) {
-        this.id = id;
+    public double[] getAverageRatings() {
+        Map<UUID, Integer> rating;
+        if (this.getSettings().getRatings() != null) {
+            rating = this.getSettings().getRatings();
+        } else if (Settings.Enabled_Components.RATING_CACHE) {
+            rating = new HashMap<>();
+        } else {
+            rating = DBFunc.getRatings(this);
+        }
+        int size = 1;
+        if (!Settings.Ratings.CATEGORIES.isEmpty()) {
+            size = Math.max(1, Settings.Ratings.CATEGORIES.size());
+        }
+        double[] ratings = new double[size];
+        if (rating == null || rating.isEmpty()) {
+            return ratings;
+        }
+        for (Entry<UUID, Integer> entry : rating.entrySet()) {
+            int current = entry.getValue();
+            if (Settings.Ratings.CATEGORIES.isEmpty()) {
+                ratings[0] += current;
+            } else {
+                for (int i = 0; i < Settings.Ratings.CATEGORIES.size(); i++) {
+                    ratings[i] += current % 10 - 1;
+                    current /= 10;
+                }
+            }
+        }
+        for (int i = 0; i < size; i++) {
+            ratings[i] /= rating.size();
+        }
+        return ratings;
     }
 
 }
