+++ Core/src/main/java/com/plotsquared/core/plot/PlotCluster.java
@@ -25,15 +25,15 @@
  */
 package com.plotsquared.core.plot;
 
+import com.plotsquared.core.PlotSquared;
 import com.plotsquared.core.database.DBFunc;
 import com.plotsquared.core.location.BlockLoc;
 import com.plotsquared.core.location.Location;
-import com.plotsquared.core.util.MainUtil;
 import com.plotsquared.core.util.RegionUtil;
 import com.sk89q.worldedit.regions.CuboidRegion;
+
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-
 import java.util.HashSet;
 import java.util.UUID;
 import java.util.function.Consumer;
@@ -156,13 +156,16 @@ public class PlotCluster {
     public void getHome(@Nonnull final Consumer<Location> result) {
         final BlockLoc home = this.settings.getPosition();
         Consumer<Location> locationConsumer = toReturn ->
-            MainUtil.getHighestBlock(this.area.getWorldName(), toReturn.getX(), toReturn.getZ(),
-                max -> {
-                    if (max > toReturn.getY()) {
-                        result.accept(toReturn.withY(1 + max));
-                    } else {
-                        result.accept(toReturn);
-                    }
+            PlotSquared.platform().getWorldUtil().getHighestBlock(this.area.getWorldName(), toReturn.getX(), toReturn.getZ(),
+                highest -> {
+                if (highest == 0) {
+                    highest = 63;
+                }
+                if (highest > toReturn.getY()) {
+                    result.accept(toReturn.withY(1 + highest));
+                } else {
+                    result.accept(toReturn);
+                }
             });
         if (home.getY() == 0) {
             // default pos
