+++ Core/src/main/java/com/plotsquared/core/util/HashUtil.java
@@ -1,0 +1,57 @@
+/*
+ *       _____  _       _    _____                                _
+ *      |  __ \| |     | |  / ____|                              | |
+ *      | |__) | | ___ | |_| (___   __ _ _   _  __ _ _ __ ___  __| |
+ *      |  ___/| |/ _ \| __|\___ \ / _` | | | |/ _` | '__/ _ \/ _` |
+ *      | |    | | (_) | |_ ____) | (_| | |_| | (_| | | |  __/ (_| |
+ *      |_|    |_|\___/ \__|_____/ \__, |\__,_|\__,_|_|  \___|\__,_|
+ *                                    | |
+ *                                    |_|
+ *            PlotSquared plot management system for Minecraft
+ *                  Copyright (C) 2020 IntellectualSites
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.plotsquared.core.util;
+
+import javax.annotation.Nonnull;
+
+public final class HashUtil {
+
+    private HashUtil() {
+    }
+
+    /**
+     * Hashcode of a boolean array.<br>
+     * - Used for traversing mega plots quickly.
+     *
+     * @param array Booleans to hash
+     * @return hashcode
+     */
+    public static int hash(@Nonnull final boolean[] array) {
+        if (array.length == 4) {
+            if (!array[0] && !array[1] && !array[2] && !array[3]) {
+                return 0;
+            }
+            return ((array[0] ? 1 : 0) << 3) + ((array[1] ? 1 : 0) << 2) + ((array[2] ? 1 : 0) << 1)
+                + (array[3] ? 1 : 0);
+        }
+        int n = 0;
+        for (boolean anArray : array) {
+            n = (n << 1) + (anArray ? 1 : 0);
+        }
+        return n;
+    }
+
+}
